Fix: accept hex/octal/binary integer literals > INT64_MAX (e.g. 0x9e3779b185ebca87)
Use stoull+static_cast<int64_t> instead of stoll for hex, octal, and binary
integer literals so full 64-bit unsigned bit patterns are accepted. Decimal
literals retain the signed range check (writing 18446744073709551615 in
decimal is almost always a mistake). All 369 tests pass.

Agent-Logs-Url: https://github.com/RyAnPr1Me/omscript/sessions/356a3860-aa0a-436a-af4d-564508ca3c51

+++ src/lexer.cpp
@@ -191,7 +191,7 @@ Token Lexer::scanNumber() {
             }
             Token token = makeToken(TokenType::INTEGER, num);
             try {
-                token.intValue = std::stoll(num, nullptr, 16);
+                token.intValue = static_cast<int64_t>(std::stoull(num, nullptr, 16));
             } catch (const std::out_of_range&) {
                 lexError("Integer literal out of range: " + num, token.line, token.column);
             }
@@ -210,7 +210,7 @@ Token Lexer::scanNumber() {
             }
             Token token = makeToken(TokenType::INTEGER, num);
             try {
-                token.intValue = std::stoll(num.substr(2), nullptr, 8);
+                token.intValue = static_cast<int64_t>(std::stoull(num.substr(2), nullptr, 8));
             } catch (const std::out_of_range&) {
                 lexError("Integer literal out of range: " + num, token.line, token.column);
             }
@@ -229,7 +229,7 @@ Token Lexer::scanNumber() {
             }
             Token token = makeToken(TokenType::INTEGER, num);
             try {
-                token.intValue = std::stoll(num.substr(2), nullptr, 2);
+                token.intValue = static_cast<int64_t>(std::stoull(num.substr(2), nullptr, 2));
             } catch (const std::out_of_range&) {
                 lexError("Integer literal out of range: " + num, token.line, token.column);
             }
